// old/Oldfrontend/src/Pages/Dashboard/component/SystemIdleGame/types.ts
export type ToolType = 'empty' | 'water' | 'axe' | 'scythe';
// ...
export type Direction = 'down' | 'up' | 'left' | 'right';